mood-ring/requirements.txt
textblob
customtkinter
pillow
